# Make hassfest.dependencies faster with multiprocessing