@balloob 2022-11-03
Current version of this check runs in 4.5s on my M1 macbook.

Instead of wrapping it all in a pool, it would be better to find what is so slow on your end. Would it be the check if a file/dir exists:

https://github.com/akx/home-assistant-core/blob/faster-hassfest-deps/script/hassfest/dependencies.py#L211-L213

And if so, could that be done in a single fetch of the whole directly listing instead of doing individual checks.