@akx 2022-11-03
@balloob Lucky you – I'm still running a 2019 clunker with "only" a 2,4 GHz 8-Core Core i9... 😁 

I did run a profiler, of course, before embarking on optimizing things. By far the most time spent (in a profiler run, so the times shown in the screenshot are inflated), is the `Collector.collect` hierarchy that does `parse` for each file and a whole lot of AST visitation.

This screenshot is from the PyCharm profiler's output at 9c85d22bab333556efd4194fc74842b044412cfe.

![home-assistant-core14](https://user-images.githubusercontent.com/58669/199785814-762eb2c0-5fc2-4f2e-bdd5-4f5350cf0122.png)

(There's no point comparing the same diagram post-multiprocessing-enablement, since the child processes aren't slowed down by the profiler.)