@akx 2022-11-03
@balloob Ah, actually the https://github.com/akx/home-assistant-core/blob/faster-hassfest-deps/script/hassfest/dependencies.py#L211-L213 branch is never even taken even on `dev` (see code coverage gutter being red), so that won't be a source of slowness :)

<img width="704" alt="Screenshot 2022-11-03 at 19 16 28" src="https://user-images.githubusercontent.com/58669/199790219-a5a5b1ec-ffbd-4fad-be60-7d9a0bdc2191.png">