@akx 2022-11-16
@balloob Have you had a chance to look at the discussion above? I rebased this branch, and the perf increase seems relevant still.

In GitHub Actions, the `Check hassfest` action says
* [`dev`](https://github.com/home-assistant/core/actions/runs/3479878171/jobs/5818987578): `Validating dependencies... done in 12.41s`
* this PR: `Validating dependencies... done in 5.41s`

Locally: 

```
$ git checkout dev && hyperfine 'python -m script.hassfest -p dependencies' && git checkout faster-hassfest-deps && Switched to branch 'dev'
Your branch is up to date with 'origin/dev'.
hyperfine 'python -m script.hassfest -p dependencies'
Benchmark 1: python -m script.hassfest -p dependencies
  Time (mean ± σ):     14.578 s ±  0.846 s    [User: 13.019 s, System: 0.951 s]
  Range (min … max):   13.863 s … 16.779 s    10 runs

Switched to branch 'faster-hassfest-deps'
Your branch is up to date with 'origin/faster-hassfest-deps'.
Benchmark 1: python -m script.hassfest -p dependencies
  Time (mean ± σ):      4.612 s ±  0.172 s    [User: 36.067 s, System: 5.783 s]
  Range (min … max):    4.418 s …  4.925 s    10 runs
$
```