@akx 2023-01-23
Rebased once again, on top of a1a324a02eeee25bfc81309004bbbefa99aab0c3 now 😄 

This is still ~3x faster than `dev` on my box.

```
$ git checkout dev && hyperfine 'python -m script.hassfest -p dependencies' && git checkout faster-hassfest-deps && hyperfine 'python -m script.hassfest -p dependencies'
Switched to branch 'dev'
Your branch is up to date with 'upstream/dev'.
Benchmark 1: python -m script.hassfest -p dependencies
  Time (mean ± σ):      8.683 s ±  0.424 s    [User: 8.138 s, System: 0.477 s]
  Range (min … max):    8.395 s …  9.535 s    10 runs

Switched to branch 'faster-hassfest-deps'
Your branch is up to date with 'origin/faster-hassfest-deps'.
Benchmark 1: python -m script.hassfest -p dependencies
  Time (mean ± σ):      2.969 s ±  0.153 s    [User: 25.163 s, System: 2.940 s]
  Range (min … max):    2.766 s …  3.183 s    10 runs
```